feat(linting rules): update stylelint configuration
The tests are adapted as the rule that was previously used is removed from the linting
configuration. The version is set to 1.0.0 to prepare the release of a new version of this package.
Also the CHANGELOG.md is prepared for release with detailed information what changes in this
release.

BREAKING CHANGE: Many linting rules are added, removed or updated. Support for styleint versions
lower than 14.0.0 is dropped.