feat(index.js): disable no-empty-source rule
Concerning anuglar, when creating a component one typically adds three files, including a *.component.scss file for style. Even if the component does not has any styles, the file should be commited and configuration (link in @component annotation) should be added to be prepared if styles must be added in the future.